+++ bauh/__init__.py
@@ -1,4 +1,4 @@
-__version__ = '0.9.29'
+__version__ = '0.10.0'
 __app_name__ = 'bauh'
 
 import os