+++ bauh/api/abstract/context.py
@@ -14,7 +14,7 @@ class ApplicationContext:
     def __init__(self, download_icons: bool, http_client: HttpClient, app_root_dir: str, i18n: I18n,
                  cache_factory: MemoryCacheFactory, disk_loader_factory: DiskCacheLoaderFactory,
                  logger: logging.Logger, file_downloader: FileDownloader, distro: str, app_name: str,
-                 internet_checker: InternetChecker, root_user: bool):
+                 internet_checker: InternetChecker, root_user: bool, screen_width: int = -1, screen_height: int = -1):
         """
         :param download_icons: if packages icons should be downloaded
         :param http_client: a shared instance of http client
@@ -27,6 +27,8 @@ class ApplicationContext:
         :param distro
         :param app_name
         :param internet_checker
+        :param screen_width
+        :param screen_height
         """
         self.download_icons = download_icons
         self.http_client = http_client
@@ -44,6 +46,8 @@ class ApplicationContext:
         self.root_user = root_user
         self.root_password = None
         self.internet_checker = internet_checker
+        self.screen_width = screen_width
+        self.screen_height = screen_height
 
     def is_system_x86_64(self):
         return self.arch_x86_64